requirements.txt
imgaug==0.2.7
numpy==1.17.0
opencv-python==4.0.0.21
pandas==0.23.4
Pillow==5.4.1
scikit-learn==0.21.3
tensorflow-gpu==1.13.1
urllib3==1.24.1
